fix(build): ignore semantic-release errors
semantic-release will exit with non-zero exit codes if there is no new
version to release, as such, any `chore`/`style`/`doc` only commits will
fail the build. This change makes it so that the exit code semantic
release emits is ignored, and the build will pass, even if it fails.